로그인 인증 상태 유지 및 로그아웃 기능 구현

+++ Chapter06/src/main/java/com/springbootTest/controller/BoardController.java
@@ -8,19 +8,27 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import com.springbootTest.domain.Board;
+import com.springbootTest.domain.Member;
 import com.springbootTest.service.BoardService;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 	@Autowired
 	private BoardService boardService;
 	
 	@RequestMapping("/getBoardList")
-	public String getBoardList(Model model, Board board) {
+	public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		List<Board> boardList = boardService.getBoardList(board);
 		
 		model.addAttribute("boardList", boardList);
@@ -28,30 +36,50 @@ public class BoardController {
 	}
 	
 	@GetMapping("/insertBoard")
-	public String inserBoardView() {
+	public String inserBoardView(@ModelAttribute("member") Member member) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		return "insertBoard";
 	}
 	
 	@PostMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		boardService.insertBoard(board);
 		return "redirect:getBoardList";
 	}
 	
 	@GetMapping("/getBoard")
-	public String getBoard(Board board, Model model) {
+	public String getBoard(@ModelAttribute("member") Member member, Board board, Model model) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		model.addAttribute("board", boardService.getBoard(board));
 		return "getBoard";
 	}
 	
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		boardService.updateBoard(board);
 		return "forward:getBoardList";
 	}
 	
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+		
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 	}
@@ -59,6 +87,11 @@ public class BoardController {
 	@GetMapping("/hello")
 	public void hello(Model model) {
 		model.addAttribute("greeting", "Hello 타임리프.^^");
+	}
+	
+	@ModelAttribute("member")
+	public Member setMember() {
+		return new Member();
 	}
